QBS: Set chart maximum to a fixed 15.

+++ QBS/include/QBS_charth.h
@@ -36,7 +36,7 @@ protected:
 
     void draw_y_axis();
 
-    int max_;
+    double max_;
 
 };
